\name{OIEmapper}
\alias{OIEmapper}
\title{
Generate KML file for certain disease.
}
\description{
}
\usage{
OIEmapper(disease, dbfile)
}
\arguments{
  \item{disease}{
}
  \item{dbfile}{
}
}
\details{
}
\value{

}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{
\dontrun{
OIEmapper(disease='African swine fever', dbfile='OIEdb.sqlite')
}
}
