\name{OIEparser-package}
\alias{OIEparser-package}
\docType{package}
\title{
\packageTitle{OIEparser}
}
\description{
\packageDescription{OIEparser}
}
\details{
\packageDESCRIPTION{OIEparser}
\packageIndices{OIEparser}
}
\author{
\packageAuthor{OIEparser}

Maintainer: \packageMaintainer{OIEparser}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
