\name{OIEparser-package}
\alias{OIEparser-package}
\alias{OIEparser}
\docType{package}
\title{
\packageTitle{OIEparser}
}
\description{
\packageDescription{OIEparser}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OIEparser}
\packageIndices{OIEparser}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{OIEparser}

Maintainer: \packageMaintainer{OIEparser}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
